requirements.txt
openai-whisper
rich
questionary
